feat: add utils and arrow foldders
